requirements.txt
# Core LangChain
langchain>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0
langchain-huggingface>=0.1.2
langchain-chroma>=0.1.4

# Vector Store
chromadb>=0.5.0

# Embeddings & Reranking
torch==2.2.2
numpy<2.0
transformers>=4.34.0,<5.0.0
sentence-transformers>=3.0.0,<5.0.0

# BM25 keyword search
rank_bm25>=0.2.2

# Evaluation
ragas>=0.2.0
datasets>=2.20.0

# Document Loading
pypdf>=4.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
tiktoken>=0.7.0
